refactor: Enrollment 관련 exception 메서드들을 global에서 domain/enrollment/exception으로 이동
